Remove unnecessary u64 cast in canonical epoch computation
parse_yyyymmdd_ascii already returns u64, so `days as u64` is a no-op
cast that newer clippy flags under unnecessary_cast. The other three
casts (hour, min, sec) are from u32 and must stay.

+++ crates/auth/src/canonical.rs
@@ -365,7 +365,7 @@ pub fn parse_amz_date(ts: &str) -> Option<u64> {
         return None;
     }
 
-    Some(days as u64 * 86400 + hour as u64 * 3600 + min as u64 * 60 + sec as u64)
+    Some(days * 86400 + hour as u64 * 3600 + min as u64 * 60 + sec as u64)
 }
 
 pub(crate) fn parse_amz_date_stamp(date_stamp: &str) -> Option<u64> {
